# api/openapi.yml
openapi: 3.0.3
info:
  title: Manjo Ticketing API
  version: 1.0.0
  description: |
    Complete HTTP surface of the Go API (`backend/`).

    ## The envelope

    Every JSON endpoint answers `{ "code": 200000, "message": "Success", "data": ... }` on success
    and `{ "code": <error code>, "message": ..., "data": ... }` on failure — the HTTP status line
    and the envelope code both carry meaning, and a client that branches only on status will miss
    distinctions the numeric code makes.

    Two endpoints sit outside it: `GET /healthz` (bare JSON, read by orchestrators) and
    `GET /api/v1/ticket/order/{order_id}/qris.png` (a PNG).

    ## Identifiers

    Public routes address an **event by its slug** and an **order by its order number**. Admin
    routes address everything by **UUID**. The two namespaces are not interchangeable — an order
    has both a public order number and an admin-side UUID, and passing one where the other is
    expected is a 400 or a 404, never a silent success.

    ## Authentication

    `/api/v1/admin/*` requires a bearer JWT from `POST /api/v1/admin/login`. `POST /v1.0/callback/exec`
    is the payment gateway's own endpoint and authenticates with a *different* bearer credential
    (`PG_CALLBACK_TOKEN`). Everything else is deliberately unauthenticated — the guest purchase
    journey has no accounts (Constitution Principle VI).

    ## Rate limits

...
    per-IP budget, and the SSE status stream caps concurrent connections per IP rather than
...

...
servers:
...
security: []
...
paths:
...
  /api/v1/ticket/{slug}:
    get:
      tags: [Public - Catalogue]
      summary: List ticket types for an event
      operationId: listTicketTypes
      parameters:
        - $ref: '#/components/parameters/EventSlug'
      description: |
        Sellable ticket types with **remaining** quota and the per-ticket admission window
        (`event_start` / `event_end`), which is the ticket's own window and not the parent event's
        dates.
      responses:
        '200':
          description: Ticket types on sale.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TicketTypeSummary'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/admin/events:
    get:
      tags: [Admin - Events]
      summary: List events
      operationId: adminListEvents
      security: [{ adminAuth: [] }]
      description: |
        Every event regardless of status, unlike the public list — one page at a time.

        This is the operator's browsing table. It is NOT the read a filter dropdown should use:
        one page of it would name only the first events and silently hide the rest. Use
        `GET /api/v1/admin/events/options` for selectors.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: One page of events, newest start date first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageMeta'
                          - type: object
                            required: [items]
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/EventAdminView'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Admin - Events]
      summary: Create an event
      operationId: adminCreateEvent
      security: [{ adminAuth: [] }]
      description: |
        `slug` is unique and must be lowercase letters, digits and single hyphens. Uniqueness is
        decided by the constraint rather than a pre-check, because nothing else can decide it
        without a race.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventAdminView'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: The slug is already taken.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/events/options:
    get:
      tags: [Admin - Events]
      summary: Every event as an id/name pair, for filter selectors
      operationId: adminListEventOptions
      security: [{ adminAuth: [] }]
      description: |
        The full catalogue in selector form, ordered by name. Deliberately **not** paginated: the
        order and attendee lists filter by event, and a dropdown fed from one page of
        `GET /api/v1/admin/events` would offer only the first events with nothing to indicate the
        rest existed. Two columns is what keeps staying unpaginated cheap.

        Route order matters in the implementation: this must be registered before
        `/api/v1/admin/events/{id}`, or `options` is parsed as an event UUID.
      responses:
        '200':
          description: Every event, as id and name.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/EventOption'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/admin/events/{id}:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Admin - Events]
      summary: Get an event
      operationId: adminGetEvent
      security: [{ adminAuth: [] }]
      description: Full detail, including the event's ticket types with their `sold` counts.
      responses:
        '200':
          description: The event.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventAdminDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Admin - Events]
      summary: Update an event
      operationId: adminUpdateEvent
      security: [{ adminAuth: [] }]
      description: Full replacement — send every field. Invalidates the cached public lists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventAdminView'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The slug is already taken.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Admin - Events]
      summary: Delete an event
      operationId: adminDeleteEvent
      security: [{ adminAuth: [] }]
      description: Guarded — an event with orders against it cannot be deleted.
      responses:
        '204':
          description: Deleted. No body.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The event has orders against it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/ticket-types:
    get:
      tags: [Admin - Ticket Types]
      summary: List ticket types
      operationId: adminListTicketTypes
      security: [{ adminAuth: [] }]
      description: '`event_id` is required — there is no unscoped listing.'
      parameters:
        - name: event_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The event's ticket types.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TicketTypeAdminView'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Admin - Ticket Types]
      summary: Create a ticket type
      operationId: adminCreateTicketType
      security: [{ adminAuth: [] }]
      description: |
        `quota` on create is the allocation; from then on the stored value is what **remains**.

        `event_start` / `event_end` bound when the ticket **admits**, not showtime — validation
        admits no tolerance, so an admin who sets `event_start` to showtime turns away every early
        arrival. The window must sit inside the parent event's dates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketTypeRequest'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TicketTypeAdminView'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/ticket-types/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [Admin - Ticket Types]
      summary: Get a ticket type
      operationId: adminGetTicketType
      security: [{ adminAuth: [] }]
      responses:
        '200':
          description: The ticket type.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TicketTypeAdminView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Admin - Ticket Types]
      summary: Update a ticket type
      operationId: adminUpdateTicketType
      security: [{ adminAuth: [] }]
      description: |
        Full replacement. `event_id` is ignored — a ticket type does not move between events.
        Writing `quota` sets the **remaining** count, not an allocation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketTypeRequest'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TicketTypeAdminView'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin - Ticket Types]
      summary: Delete a ticket type
      operationId: adminDeleteTicketType
      security: [{ adminAuth: [] }]
      description: Guarded against deleting a type that has been sold or that a package draws on.
      responses:
        '204':
          description: Deleted. No body.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The ticket type has been sold, or a package depends on it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The token from `POST /api/v1/admin/login`. HS256, verified against `JWT_SECRET`; only
        HS256 is accepted, because letting the token's own header select the algorithm is what
        makes algorithm-confusion forgeries possible.
    gatewayCallbackAuth:
      type: http
      scheme: bearer
      description: |
        `PG_CALLBACK_TOKEN` — the payment gateway's credential, entirely separate from the admin
        JWT. Compared in constant time, and unconditionally: returning early on the first differing
        byte would leak the expected token a character at a time to anyone able to time the
        endpoint.
  parameters:
    EventSlug:
      name: slug
      in: path
      required: true
      description: The event's public slug.
      schema:
        type: string
      example: jazz-night-2026
    EventId:
      name: id
      in: path
      required: true
      description: Event UUID.
      schema:
        type: string
        format: uuid
    ResourceId:
      name: id
      in: path
      required: true
      description: The resource's UUID.
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      description: |
        1-based page number. Corrected rather than refused: zero, negative, non-numeric and
        absent all resolve to page 1, and a page past the last resolves to the last page. Read
        the returned `page` to see which one was actually served — a stale bookmark self-corrects
        instead of erroring.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      description: |
        Rows per page. Values above the maximum are reduced to it; values below 1, non-numeric
        values and absent all take the default. Never refused. The cap is enforced server-side:
        no request can demand an unbounded page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    ValidationError:
      description: |
        The request was malformed. `data` carries a field map when more than one field is at fault.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400001
            message: Some fields are missing or invalid.
            data:
              attendees[0].phone: Phone number must be 12-15 digits.
...
    Unauthorized:
      description: No bearer token, or the token is invalid or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401001
            message: A bearer token is required.
            data: null
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404001
            message: Event not found.
            data: null
...
  schemas:
    PageMeta:
      type: object
      description: |
        Where a page sits in its list. Carried alongside `items` in `data` on every paginated
        admin list. `page` and `page_size` are what the server SERVED, which may differ from
        what was requested — see the `page` parameter.
      required: [page, page_size, total, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 2
        page_size:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          format: int64
          minimum: 0
          description: Rows matching the current filters, ignoring paging.
          example: 137
        total_pages:
          type: integer
          minimum: 0
          description: Zero when `total` is zero — never "page 1 of 0".
          example: 7
    EventOption:
      type: object
      description: An event reduced to what a filter dropdown needs.
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Jazz Night 2026
    Envelope:
      type: object
      description: |
        The one JSON shape every endpoint returns. `code` is 200000 on success regardless of which
        2xx the status line carries.
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 200000
        message:
          type: string
          example: Success
        data:
          nullable: true
          description: The typed payload, or null.
    Error:
      type: object
      description: |
        The failure envelope. `data` is null except where a contract attaches detail — a validation
        field map, the current QR payload on PAYMENT_ALREADY_STARTED, the remaining cooldown on a
        guest resend refusal.
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: |
            The numeric envelope code. Note that several distinct conditions render as 400001, so
            a client that must tell them apart reads the stable string code where one is offered
            (see `AvailabilityReason.code`).
          example: 400001
        message:
          type: string
        data:
          nullable: true
    Money:
      type: string
      description: |
        A decimal amount as a string, so no precision is lost in transit. Responses are always
        strings; requests also accept a JSON number.
      example: "150000.00"

    # --- Public catalogue ---------------------------------------------------
...
    TicketTypeSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description:
          type: string
          nullable: true
          description: |
            Replaces the booking card's standard non-refundable notice for this ticket. Null keeps
            the standard wording.
        price: { $ref: '#/components/schemas/Money' }
        quota_remaining:
          type: integer
          format: int32
          description: Seats still available. Never an original allocation.
        sales_start: { type: string, format: date-time }
        sales_end: { type: string, format: date-time }
        event_start:
          type: string
          format: date-time
          description: |
            When admission OPENS for this ticket — the ticket type's own window, not the parent
            event's dates.
        event_end: { type: string, format: date-time }
...
    EventRequest:
      type: object
      required: [name, slug, venue, address, start_date, end_date, status]
      properties:
        name: { type: string }
        slug:
          type: string
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
          description: Lowercase letters, digits and single hyphens. Unique.
          example: jazz-night-2026
        description: { type: string, nullable: true }
        venue: { type: string }
        address: { type: string }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        banner_url: { type: string, nullable: true }
        status: { type: string, enum: [DRAFT, PUBLISHED] }
        scale:
          type: integer
          format: int64
          nullable: true
          description: Expected visitor count. Null clears it.
    EventAdminView:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        venue: { type: string }
        address: { type: string }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        banner_url: { type: string, nullable: true }
        status: { type: string, enum: [DRAFT, PUBLISHED] }
        scale: { type: integer, format: int64, nullable: true }
        created_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time, nullable: true }
    EventAdminDetail:
      allOf:
        - $ref: '#/components/schemas/EventAdminView'
        - type: object
          properties:
            ticket_types:
              type: array
              items: { $ref: '#/components/schemas/TicketTypeAdminView' }
    TicketTypeRequest:
      type: object
      required: [name, price, quota, sales_start, sales_end, event_start, event_end]
      properties:
        event_id:
          type: string
          format: uuid
          description: Required on create; ignored on update — a ticket type does not move events.
        name: { type: string }
        description: { type: string, nullable: true }
        price: { $ref: '#/components/schemas/Money' }
        quota:
          type: integer
          format: int32
          minimum: 0
          description: |
            The allocation on create; the **remaining** count on every write thereafter. This field
            is never an original allocation once the ticket type exists.
        sales_start: { type: string, format: date-time }
        sales_end: { type: string, format: date-time }
        event_start:
          type: string
          format: date-time
          description: |
            When admission OPENS, not showtime. Validation admits no tolerance, so setting this to
            showtime turns away every early arrival. Must sit inside the parent event's dates.
        event_end: { type: string, format: date-time }
    TicketTypeAdminView:
      type: object
      properties:
        id: { type: string, format: uuid }
        event_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        price: { $ref: '#/components/schemas/Money' }
        quota:
          type: integer
          format: int32
          description: The count that REMAINS, never an original allocation.
        sold: { type: integer }
        sales_start: { type: string, format: date-time }
        sales_end: { type: string, format: date-time }
        event_start: { type: string, format: date-time }
        event_end: { type: string, format: date-time }
...